Refactor CSS properties to use dense IDs and bitsets:
Introduce contiguous PropertyId and ShorthandId enums backed by name-to-ID maps, replacing hash-valued property identifiers throughout the style system.
Replace PropertyIdSet's unordered set with a fixed-size bitset, eliminating heap allocations during style state changes while providing deterministic ascending iteration.
Keep name hashes for stylesheet storage and unknown or custom properties, while using dense typed IDs for registered-property lookup and dispatch.
Preserve property aliases and the public Checked compatibility identifier.
Add transaction-end reconciliation for interdependent image sizing properties, deferring percentage-based calculations until their containing block is resolved.
Keep formatting-context-owned sizes intact for table cells and inline content.
Add validation and tests for dense built-in registration, custom IDs, aliases, bitset operations, allocation-free storage, unknown properties, property application order, and image sizing regressions.

+++ include/eepp/ui/models/csspropertiesmodel.hpp
@@ -64,7 +64,7 @@ class CSSPropertiesModel final : public Model {
 		mWidget = widget;
 		mData = widget ? widget->getPropertiesImplemented() : std::vector<PropertyId>();
 		for ( const auto& prop : mData ) {
-			const auto* def = StyleSheetSpecification::instance()->getProperty( (Uint32)prop );
+			const auto* def = StyleSheetSpecification::instance()->getProperty( prop );
 			if ( !def )
 				continue;
 			mProps[prop] = def;
